Add XC potential component calculation routine
For linear response the XC potential is computed by multiplying the
second derivative of the XC functional with the perturbed density. For
the spin-separated case, this involves four components
(aa*a, ab*b, ba*a, bb*b)
which should be added two and two and then multiplied to the orbitals
(aa*a*a + ab*b*a and ba*b*b + bb*b*b).

Each component is assembled identically in the same routine.