Coulomb and exchange operator tests
The reference values of these tests are obtained by a standard run.
The result is then checked to a much higher precision than the actual
calculation. The values are therefore not correct to that many digits,
but the test will be able to detect small changes in how the
calculation is performed. This means however that some legitimate
changes in the algorithm might result in a change of the reference
value within the requested precision, which will however cause the
tests to fail.