Update on GGA gamma vs explicit.
Not yes sure which of the two implementations is right. Preliminary
tests showed that gamma-version agreed better with FD (see previous
commit), whereas at high precision (eps=1.0e-6) the outcome seems
reversed. I also noticed that the way grids are built in the XC
potential calculations has a large impact. For example, when grids for
multiplications and function scalar products (between two
FunctionTreeVector-s) are borrowed from the densities and their
gradients the result is affected up to a few percent.

Gamma: 6.9003819475
FD   : 6.92776      (5p stencil with E=0.001 and E=0.002)
Expl : 6.9286103891