Found the last mistake with the gamma-type calculation.
I was simply including one contributuon twice and forgetting one.
Additionally now both gamma and explicit gradient calculations are
performed using the very same grid, which is the union of the ground state density
grid and the perturbed density grid. There is only one step where the
grid is different: the last contribution in the gamma-type derivative
(the one with the first derivative wrt gamma times grad gamma). I
doubt that has a large impact and I'm not sure we should change it.