Debugging the node-wise calculation of the XC potentials.
1) XC first derivative unit tests are now passing
2) XC second derivative unit tests are running through but failing
3) For the xc_hessian test (LDA) I suspect it is only a question of
accuracy, because the numerical differences are minimal
4) For the xc_hessian_pbe the issue might be more severe because the
differences are more pronounced

One crucial question is the accuracy used to compute the densities. In
the previous implementation all densities were computed up to a given
accuracy and then used as functions in the subsequento calculations of
the potentials. In the current implementation, however, we need the
same grid for all densities, otherwise the calculation of the
potentials which now happens at node level (no longer at function
level) would not be possible. The way this is achieved for now is by a
refine_grid pass for the perturbed density. This might be problematic
if the perturbed density is somewhere more refined than the GS
density.