Towards GGA Linear response
A couple more bug fixes along the way of getting GGA response running
for closed and open shell alike with the semicontracted implementation
n XCFunctional.

In all calculations we now keep the energy density among the output
functions. This is not strictly necessary for response calculations,
however it simplifies the implementation to keep it always and it
should not be too large an overhead. If we need to save memory, it is
however simple to just delete the corresponding FunctionTree when it
is not needed.