Fix bug in calcExchange
coef_vec was initialized outside the loop whereas finc_vec was
populated inside. This led to a mismatch in position of functions and
respective coefficients

index in getSpinFactor was switched. Now it should be right, although
for all practical cases it should not matter, as long as we do not
allow paired and unpaired electrons at the same time (e.g. ROHF).